Splitting up code into separate files, might be easier for testing down the line
