+++ service.go
@@ -1,16 +1,10 @@
 package main
 
 import (
-	"encoding/json"
-	"errors"
-	"fmt"
-	"io"
 	"log"
 	"net/http"
 	"os"
 	"path/filepath"
-	"regexp"
-	"strings"
 	"time"
 )
 
@@ -28,118 +22,61 @@ func IsOnline() bool {
 	return true
 }
 
-func GetIanaName() (string, error) {
-    linkPath := "/etc/localtime"
-    targetPath, err := os.Readlink(linkPath)
-    if err != nil {
-        return "", err
-    }
-
-    tzParts := strings.Split(targetPath, "/")
-    if len(tzParts) < 3 {
-        return "", errors.New("invalid timezone format")
-    }
-
-    continent, country := tzParts[len(tzParts)-2], tzParts[len(tzParts)-1]
-    timezone := fmt.Sprintf("%s/%s", continent, country)
-
-    _, err = time.LoadLocation(timezone)
-    if err != nil {
-        return "", err
-    }
-
-    return timezone, nil
-}
-
-func ValidateIanaName(timezone string) (bool, error) {
-
-	validTimezoneRegex := `^[A-Za-z]+(/[A-Za-z_-]+)+$`
-    matched, err := regexp.MatchString(validTimezoneRegex, timezone)
-
-	if err != nil {
-        return false, fmt.Errorf("Error compiling regex: %v", err)
-    }
-
-    if !matched {
-        return false, fmt.Errorf("Invalid timezone format: %s", timezone)
-    }
-
-	return true, nil
-}
-
 func execTzChange(){
 	timezoneDir := "/usr/share/zoneinfo"
-    // localtimePath := "/etc/localtime"
-	sysTz, err := GetIanaName()
+    localtimePath := "/etc/localtime"
+
+	sysTz, err := GetLocalIanaName()
 
 	if err != nil {
 		log.Fatal(err)
 		return
 	}
 
-	ipapiClient := http.Client{}
-
-	req, err := http.NewRequest("GET", "https://ipapi.co/json/", nil)
-
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	req.Header.Set("User-Agent", "ipapi.co/#go-v1.3")
-
-	resp, err := ipapiClient.Do(req)
-
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	defer resp.Body.Close()
-
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		log.Fatal(err)
-	}
+	ipGeoTz := GetRemoteIanaName()
 	
-	var inputJson map[string]interface{}
-
-	json.Unmarshal([]byte(string(body)), &inputJson)
-	
-	ipGeoTz := inputJson["timezone"].(string)
-	tempTz := "America/Los_Angeles"
 	// Checking to make sure that someone isn't trying to pass malicious code in the ipGeoTz string
-	isValid, err := ValidateIanaName(tempTz)
+	isValid, err := ValidateIanaName(ipGeoTz)
 
 	if !isValid {
-		log.Printf("Timezone is not valid: %s", tempTz)
+		log.Fatalf("Timezone is not valid: %s", ipGeoTz)
 	} else if (ipGeoTz != sysTz) {
 		timezonePath := filepath.Join(timezoneDir, ipGeoTz)
 
-		if _, err := os.Stat(tempTz); os.IsNotExist(err) {
-			log.Fatalf("Timezone file does not exist: %s", tempTz)
+		if _, err := os.Stat(timezonePath); os.IsNotExist(err) {
+			log.Fatalf("Timezone path does not exist: %v", err)
+			return
 		} else {
-			log.Println("Timezone is valid")
+			log.Printf("Valid Timezone: %s", ipGeoTz)
 		}
-		// else if err := os.Remove(localtimePath); err != nil {
-		// 	log.Fatal("Error removing current localtime: %v", err)
-		// } else if err := os.Symlink(timezonePath, localtimePath); err != nil {
-		// 	log.Fatal("Error creating symlink: %v", err)
-		// }
-		log.Println("Timezones do not match, changing....")
+		
+		if err := os.Remove(localtimePath); err != nil {
+			log.Fatalf("Error removing current localtime: %v", err)
+			return
+		} else if err := os.Symlink(timezonePath, localtimePath); err != nil {
+			log.Fatalf("Error creating symlink: %v", err)
+			return
+		} else {
+			log.Printf("Timezone successfully changed to: %s", ipGeoTz)
+		}
+	} else {
+		log.Println("Time zones unchanged")
 	}
 }
 
 func main() {
 	isOnline := IsOnline()
 
-	if isOnline {
-		for i := 0; i < 100; i++ {
+	if !isOnline {
+		for i := 0; i < 30; i++ {
+			log.Println("Unable to connect to internet, trying again...")
 			time.Sleep(2 * time.Second)
 			if IsOnline() == true {
 				execTzChange()
-				break
+				return
 			}
-
 		}
+		log.Fatalln("Unable to connect to the internet")
 	} else {
 		execTzChange()
 	}
